+++ cli/src/merge_tools/external.rs
@@ -20,6 +20,8 @@ use jj_lib::matchers::Matcher;
 use jj_lib::merge::Merge;
 use jj_lib::merged_tree::MergedTree;
 use jj_lib::merged_tree::MergedTreeBuilder;
+use jj_lib::repo_path::RepoPathUiConverter;
+use jj_lib::store::Store;
 use jj_lib::working_copy::CheckoutOptions;
 use pollster::FutureExt;
 use thiserror::Error;
@@ -168,12 +170,13 @@ pub enum ExternalToolError {
     Io(#[source] std::io::Error),
 }
 
-pub fn run_mergetool_external(
+fn run_mergetool_external_single_file(
     editor: &ExternalMergeTool,
-    tree: &MergedTree,
+    store: &Store,
     merge_tool_file: &MergeToolFile,
     default_conflict_marker_style: ConflictMarkerStyle,
-) -> Result<MergedTreeId, ConflictResolveError> {
+    tree_builder: &mut MergedTreeBuilder,
+) -> Result<(), ConflictResolveError> {
     let MergeToolFile {
         repo_path,
         conflict,
@@ -270,7 +273,7 @@ pub fn run_mergetool_external(
     let new_file_ids = if editor.merge_tool_edits_conflict_markers || exit_status_implies_conflict {
         conflicts::update_from_content(
             file_merge,
-            tree.store(),
+            store,
             repo_path,
             output_file_contents.as_slice(),
             conflict_marker_style,
@@ -278,8 +281,7 @@ pub fn run_mergetool_external(
         )
         .block_on()?
     } else {
-        let new_file_id = tree
-            .store()
+        let new_file_id = store
             .write_file(repo_path, &mut output_file_contents.as_slice())
             .block_on()?;
         Merge::normal(new_file_id)
@@ -302,8 +304,49 @@ pub fn run_mergetool_external(
         }),
         Err(new_file_ids) => conflict.with_new_file_ids(&new_file_ids),
     };
-    let mut tree_builder = MergedTreeBuilder::new(tree.id());
     tree_builder.set_or_remove(repo_path.to_owned(), new_tree_value);
+    Ok(())
+}
+
+pub fn run_mergetool_external(
+    ui: &Ui,
+    path_converter: &RepoPathUiConverter,
+    editor: &ExternalMergeTool,
+    tree: &MergedTree,
+    merge_tool_files: &[MergeToolFile],
+    default_conflict_marker_style: ConflictMarkerStyle,
+) -> Result<MergedTreeId, ConflictResolveError> {
+    let mut tree_builder = MergedTreeBuilder::new(tree.id());
+    for (i, merge_tool_file) in merge_tool_files.iter().enumerate() {
+        writeln!(
+            ui.status(),
+            "Resolving conflicts in: {}",
+            path_converter.format_file_path(&merge_tool_file.repo_path)
+        )?;
+        match run_mergetool_external_single_file(
+            editor,
+            tree.store(),
+            merge_tool_file,
+            default_conflict_marker_style,
+            &mut tree_builder,
+        ) {
+            Ok(()) => {}
+            Err(err) if i == 0 => {
+                // If the first resolution fails, just return the error normally
+                return Err(err);
+            }
+            Err(err) => {
+                // Some conflicts were already resolved, so we should return an error with the
+                // partially-resolved tree so that the caller can save the resolved files.
+                let resolved_tree = tree_builder.write_tree(tree.store())?;
+                return Err(ConflictResolveError::PartialResolution {
+                    source: Box::new(err),
+                    resolved_count: i,
+                    resolved_tree,
+                });
+            }
+        }
+    }
     let new_tree = tree_builder.write_tree(tree.store())?;
     Ok(new_tree)
 }