+++ CHANGELOG.md
@@ -40,6 +40,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * `jj absorb` now abandons the source commit if it becomes empty and has no
   description.
 
+* `jj resolve` will now attempt to resolve all conflicted files instead of
+  resolving the first conflicted file. To resolve a single file, pass a file
+  path to `jj resolve`.
+
 ### Deprecations
 
 * `--config-toml=TOML` is deprecated in favor of `--config=NAME=VALUE` and